requirements.txt
# Ivoris Multi-Center Pipeline

# Database
pyodbc>=5.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web API (optional)
fastapi>=0.109.0
uvicorn>=0.27.0

# Development
pytest>=7.0.0
